third_party/blink/web_tests/fast/forms/select/customizable-select/appearance-base-select-usecounter.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://issues.chromium.org/issues/361129480">
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script src="../../../../resources/testdriver.js"></script>
<script src="../../../../resources/testdriver-vendor.js"></script>

<select>
  <option>option</option>
</select>

<style>
select.button {
  appearance: base-select;
}
select.picker::picker(select) {
  appearance: base-select;
}
</style>

<script>
const select = document.querySelector('select');
const buttonUseCounterId = 5133;
const pickerUseCounterId = 5134;

promise_test(async () => {
  assert_false(internals.isUseCounted(document, buttonUseCounterId),
    'Button use counter should not be hit at the start of the test.');
  assert_false(internals.isUseCounted(document, pickerUseCounterId),
    'Picker use counter should not be hit at the start of the test.');

  select.classList.add('picker');
  assert_false(internals.isUseCounted(document, buttonUseCounterId),
    'Button use counter should not be hit with base-select only on picker.');
  assert_false(internals.isUseCounted(document, pickerUseCounterId),
    'Picker use counter should not be hit with base-select only on picker.');
  select.classList.remove('picker');

  select.classList.add('button');
  await new Promise(requestAnimationFrame);
  await new Promise(requestAnimationFrame);
  assert_true(internals.isUseCounted(document, buttonUseCounterId),
    'Button use counter should be hit with base-select only on button.');
  assert_false(internals.isUseCounted(document, pickerUseCounterId),
    'Picker use counter should not be hit with base-select only on button.');
  internals.clearUseCounter(document, buttonUseCounterId);

  select.classList.add('picker');
  await new Promise(requestAnimationFrame);
  await new Promise(requestAnimationFrame);
  assert_true(internals.isUseCounted(document, buttonUseCounterId),
    'Button use counter should be hit with base-select on button and picker.');
  assert_true(internals.isUseCounted(document, pickerUseCounterId),
    "Picker use counter should be hit if picker isn't open.");

  await test_driver.click(select);
  await new Promise(requestAnimationFrame);
  await new Promise(requestAnimationFrame);
  assert_true(internals.isUseCounted(document, pickerUseCounterId),
    'Picker use counter should be hit after picker is opened.');
}, 'UseCounters for base-select on button and picker should work.');
</script>
